Add casts to fix warnings in MSIngestonSenderTests

+++ MobileCenter/MobileCenterTests/MSIngestionSenderTests.m
@@ -73,7 +73,7 @@ static NSString *const kMSBaseUrl = @"https://test.com";
 
         XCTAssertNil(error);
         XCTAssertEqual(containerId, batchId);
-        XCTAssertEqual(statusCode, MSHTTPCodesNo200OK);
+        XCTAssertEqual((NSInteger)statusCode, MSHTTPCodesNo200OK);
 
         [expectation fulfill];
       }];
@@ -99,7 +99,7 @@ static NSString *const kMSBaseUrl = @"https://test.com";
                           NSError *error) {
 
         XCTAssertEqual(containerId, batchId);
-        XCTAssertEqual(statusCode, MSHTTPCodesNo404NotFound);
+        XCTAssertEqual((NSInteger)statusCode, MSHTTPCodesNo404NotFound);
         XCTAssertEqual(error.domain, kMSMCErrorDomain);
         XCTAssertEqual(error.code, kMSMCConnectionHttpErrorCode);
         XCTAssertEqual(error.localizedDescription, kMSMCConnectionHttpErrorDesc);
